CR031 RcsService, Enum Direction and ReadStatus
* Direction and ReadStatus move from RcsCommon to RcsService
* RcsCommon not needed, file has been removed
* Direction and Readstatus changed from int to enum

+++ core/src/com/orangelabs/rcs/service/api/ChatMessageImpl.java
@@ -25,7 +25,7 @@ public class ChatMessageImpl extends IChatMessage.Stub {
 
 	/**
 	 * Constructor
-	 * 
+	 *
 	 * @param persistentStorage ChatMessagePersistedStorageAccessor
 	 */
 	public ChatMessageImpl(ChatMessagePersistedStorageAccessor persistentStorage) {
@@ -49,7 +49,7 @@ public class ChatMessageImpl extends IChatMessage.Stub {
 	}
 
 	public int getDirection() {
-		return mPersistentStorage.getDirection();
+		return mPersistentStorage.getDirection().toInt();
 	}
 
 	public long getTimestamp() {
